Generate server group IDs with milliseconds.
Generate server group IDs with using Time.now.to_f
which gives us millisecond accuracy. This should avoid ID
collisions which could occur if two groups were spun up
in the same second thus causing image name collisions, etc on
the same machine.

+++ lib/kytoon/providers/libvirt/server_group.rb
@@ -37,7 +37,7 @@ class ServerGroup
   attr_accessor :use_sudo
 
   def initialize(options={})
-    @id = options[:id] || Time.now.to_i
+    @id = options[:id] || Time.now.to_f
     @name = options[:name]
     @use_sudo = options[:use_sudo]
     @servers=[]
